[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 83.3 MB -BaekjoonHub

+++ Java/프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int solution(float num1, float num2) {
-        return (int)((num1/num2) * 1000);
+    public int solution(int num1, int num2) {
+        return (int)(((double)num1/num2) * 1000);
     }
 }
